Literature review that was perform
This literature review was performed as a pre study for our bachelor thesis. It was conducted for the purpose of choosing the best method for the research that we were about to begin. Max and I (Karl) are publishing this for the purpose of motivating the methods that were chosen.